+++ src/common/database/migrations/1726221075535-alterEventFormAddRequireInput.ts
@@ -7,7 +7,7 @@ export class AlterEventFormAddRequireInput1726221075535
 
   public async up(queryRunner: QueryRunner): Promise<void> {
     await queryRunner.query(
-      `ALTER TABLE "event_form_option" ADD "require_input" tinyint NOT NULL CONSTRAINT "DF_05b37b136be7ec51f8054c876ed" DEFAULT 0`,
+      `ALTER TABLE "event_form_option" ADD "require_input" smallint NOT NULL CONSTRAINT "DF_05b37b136be7ec51f8054c876ed" DEFAULT 0`,
     );
     await queryRunner.query(
       `ALTER TABLE "event_form_detail" DROP COLUMN "content"`,
